Rename convert to convertCSS
