+++ src/convertCSS.ts
@@ -4,13 +4,13 @@ import { resolveColor } from '~/modules/parsed-color';
 import { ColorType } from '~/types';
 
 /**
- * Convert a color string to another format.
+ * Convert a CSS string to another format.
  *
- * @param input - The input color string.
+ * @param input - The input CSS string.
  * @param format - The output color format.
- * @returns The converted color string.
+ * @returns The converted CSS string.
  */
-export default function convert(input: string, format: ColorType): string {
+export default function convertCSS(input: string, format: ColorType): string {
   const parsed = resolveColor(input);
 
   return formatCSS(parsed[format], { format, alpha: parsed.alpha < 1 ? parsed.alpha : undefined });
